Use a different application name for each review app.
Drop the manifest.yml file as we can only define static applications there. Review instances of the apps have a dynamically generated name and therefore need to be created via the CLI.
We now use a dedicated shell script for pushing the app to CF which defines the common parameters for us.