Re-implement the concept of an "AmountSplitPersonList" in the form of a ShareList.
The new component makes heavy use of composition and delegates the actual rendering of each item to a prop function. This allows us to describe in a very DRY way how each element in the list should look like. Basic functionality and layout is thereby provided by the generic ShareList and ShareListItem component.