Report misconfigured API hosts as sentry errors
The problem with the current setup was that Raven was initialized AFTER
the api factory kicked in. Therefore, any errors generated during that
were not reported to sentry. We now load Raven in a script tag in the head
of the index.html. In order to embed the COMMIT_HASH in there, we need to
explicitly mark the index file as a template for the HtmlWebPackPlugin. For
this we also need the ejs-loader (we are using the ejs template language).
Finally, instead of disable Raven via the DEBUG switch, we use the built-in
whilelist feature.

Fixed #12.