Fix failing test
By importing the path variable from the route in the selector, we
introduced a dependency to EmptyNobtPlaceHolder component. This
component imports a picture's URL in order to display it.
Unfortunately, jest is unable to handle such imports, which
causes the test to fail. By specifying a pattern that handles
these kinds of imports (similar to the .scss imports) we avoid
processing of this file at all. Jest is also updated to v20.