Refactor environment config into separate files
The main reason for the change was to explicitly fail if one of the env
variables is not defined. However, instead of requiring them with
env-require, we now always load them from the environment and simply
fail if they are not present once we load the production-env overrides.
This is somewhat cleaner and allows us to keep things separate.