Always attach store state on error
Using the callback functionality of Raven, we can use the reference to
our store after we created it to always retrieve the latest store state
and attach it to the event that gets sent to Sentry.

Prior to this change, we would only attach the store state if an error
happened during the execution of an event (see change in
src/store/crashReporter.js). However, the current state of the store is
useful for all events.