Correctly access payload after creating nobt
We incorrectly accessed the payload of the response after creating a
nobt. This caused our reducer to not persist any data in our store which
later on caused issue #52 to happen.

In order to avoid this problem, we now pass the whole response to the
action creator and let it figure out on its own, how to access the
payload. This avoids the duplication and thereby the error.

The error was also the reason, why one always had to click twice on the
link. The first click caused a JS error and the 2nd was an actual click
on a link that made the browser perform an actual navigation (loading
the nobt again instead of using the already stored local version of it).

Fixes #52.