Rename build tasks and optimize .travis.yml
We optimize the tasks that are called in Travis to execute from
fastest to slowest.
In addition, instead of using a custom deployment provider for
uploading the sourcemaps to sentry, we use the after_deploy hook.
Reason is that a 2nd deploy provider re-triggers the before_deploy
hook which means that another build was triggered.