Update to react-router-dom v5
Version 5 ships with react hooks (f.e. useHistory) that allow
us to do all of our routing in a fairly elegant way.

To actually do the migration, several changes were necessary:

- Introduce oaf-react-router to handle scroll restoration.
Unfortunately, the UI tests didn't work with them, so I deleted them
after extensive manual testing. We are just using the libraries as
is so there is not much to test / break.

- Remove RouteGuard component.
I couldn't get the RouteGuard components to work properly with hooks
so I inlined the logic into the actual components. It is not pretty
but legacy anyway since we are moving away from redux more and more.

- Delete old snapshot unit tests.
Those are useless now that we have the strong UI test suite.

- Update cypress to version 4.